requirements.txt
# Core dependencies
langgraph==0.2.34
langchain==0.3.0
langchain-openai==0.2.0
langchain-core==0.3.0
langchain-community==0.3.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0
jupyter>=1.0.0

# Configuration and utilities
python-dotenv>=1.0.0
pyyaml>=6.0
rich>=13.0.0
typer>=0.12.0

# Database and storage
sqlite3

# DSPy for language model optimization
dspy-ai>=2.4.0

# Development dependencies
pytest>=7.0.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0
